windows/requirements.txt
pyyaml
PyMuPDF
pillow
requests
beautifulsoup4
progressbar2